refactor(plugins): remove redundant private channel runtime facades
