+++ extensions/irc/src/gateway.ts
@@ -1,10 +1,10 @@
 // Irc plugin module implements gateway behavior.
 import { runPassiveAccountLifecycle } from "openclaw/plugin-sdk/channel-outbound";
 import { createLazyRuntimeModule } from "openclaw/plugin-sdk/lazy-runtime";
+import type { RuntimeEnv } from "openclaw/plugin-sdk/runtime";
 import type { ChannelAccountSnapshot } from "openclaw/plugin-sdk/status-helpers";
 import type { ResolvedIrcAccount } from "./accounts.js";
 import { createAccountStatusSink } from "./channel-api.js";
-import type { RuntimeEnv } from "./runtime-api.js";
 import type { CoreConfig } from "./types.js";
 
 const loadIrcChannelRuntime = createLazyRuntimeModule(() => import("./channel-runtime.js"));
