--- extensions/mattermost/src/runtime-api.ts
@@ -1,65 +1,0 @@
-// Legacy map-helper exports in this facade stay for older plugin consumers.
-// New message-turn code should use createChannelHistoryWindow.
-export {
-  applyAccountNameToChannelSection,
-  applySetupAccountConfigPatch,
-  type BaseProbeResult,
-  type BlockStreamingCoalesceConfig,
-  buildChannelConfigSchema,
-  buildComputedAccountStatusSnapshot,
-  buildModelsProviderData,
-  buildPendingHistoryContextFromMap,
-  type ChannelAccountSnapshot,
-  type ChannelDirectoryEntry,
-  type ChannelGroupContext,
-  type ChannelMessageActionName,
-  type ChannelPlugin,
-  type ChatType,
-  chunkTextForOutbound,
-  clearHistoryEntriesIfEnabled,
-  createChannelHistoryWindow,
-  createAccountStatusSink,
-  createChannelPairingController,
-  createChannelMessageReplyPipeline,
-  createDedupeCache,
-  DEFAULT_ACCOUNT_ID,
-  DEFAULT_GROUP_HISTORY_LIMIT,
-  type DmPolicy,
-  formatInboundFromLabel,
-  getAgentScopedMediaLocalRoots,
-  GROUP_POLICY_BLOCKED_LABEL,
-  type GroupPolicy,
-  type HistoryEntry,
-  isDangerousNameMatchingEnabled,
-  isRequestBodyLimitError,
-  isTrustedProxyAddress,
-  listSkillCommandsForAgents,
-  loadOutboundMediaFromUrl,
-  logInboundDrop,
-  logTypingFailure,
-  migrateBaseNameToDefaultAccount,
-  type ModelsProviderData,
-  normalizeAccountId,
-  normalizeProviderId,
-  type OpenClawConfig,
-  type OpenClawPluginApi,
-  parseStrictPositiveInteger,
-  type PluginRuntime,
-  rawDataToString,
-  readRequestBodyWithLimit,
-  recordPendingHistoryEntryIfEnabled,
-  registerPluginHttpRoute,
-  type ReplyPayload,
-  resolveAllowlistMatchSimple,
-  resolveAllowlistProviderRuntimeGroupPolicy,
-  resolveChannelMediaMaxBytes,
-  resolveClientIp,
-  resolveControlCommandGate,
-  resolveDefaultGroupPolicy,
-  resolveStoredModelOverride,
-  resolveStorePath,
-  resolveThreadSessionKeys,
-  type RuntimeEnv,
-  setMattermostRuntime,
-  warnMissingProviderGroupPolicyFallbackOnce,
-} from "../runtime-api.js";
