+++ extensions/zalo/src/actions.test.ts
@@ -2,7 +2,7 @@
 import http from "node:http";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { zaloMessageActions } from "./actions.js";
-import type { OpenClawConfig } from "./runtime-api.js";
+import type { OpenClawConfig } from "./runtime-support.js";
 
 describe("zaloMessageActions.describeMessageTool", () => {
   it("honors the selected Zalo account during discovery", () => {
