--- extensions/zalo/src/runtime-api.ts
@@ -1,81 +1,0 @@
-// Zalo API module exposes the plugin public contract.
-export {
-  type BaseProbeResult,
-  type BaseTokenResolution,
-  type ChannelAccountSnapshot,
-  type ChannelMessageActionAdapter,
-  type ChannelMessageActionName,
-  type ChannelPlugin,
-  type ChannelStatusIssue,
-  type GroupPolicy,
-  type MarkdownTableMode,
-  type OpenClawConfig,
-  type OutboundReplyPayload,
-  type PluginRuntime,
-  type RegisterWebhookPluginRouteOptions,
-  type RegisterWebhookTargetOptions,
-  type ReplyPayload,
-  type RuntimeEnv,
-  type SecretInput,
-  type WizardPrompter,
-} from "./runtime-support.js";
-export {
-  DEFAULT_ACCOUNT_ID,
-  buildChannelConfigSchema,
-  createDedupeCache,
-  formatPairingApproveHint,
-  jsonResult,
-  normalizeAccountId,
-  readStringParam,
-  resolveClientIp,
-} from "./runtime-support.js";
-export {
-  addWildcardAllowFrom,
-  applyAccountNameToChannelSection,
-  applySetupAccountConfigPatch,
-  buildSingleChannelSecretPromptState,
-  mergeAllowFromEntries,
-  migrateBaseNameToDefaultAccount,
-  promptSingleChannelSecretInput,
-  runSingleChannelSecretStep,
-  setTopLevelChannelDmPolicyWithAllowFrom,
-} from "./runtime-support.js";
-export {
-  buildSecretInputSchema,
-  hasConfiguredSecretInput,
-  normalizeResolvedSecretInputString,
-  normalizeSecretInputString,
-} from "./runtime-support.js";
-export { buildTokenChannelStatusSummary, PAIRING_APPROVED_MESSAGE } from "./runtime-support.js";
-export { buildBaseAccountStatusSnapshot } from "./runtime-support.js";
-export { chunkTextForOutbound } from "./runtime-support.js";
-export { formatAllowFromLowercase, isNormalizedSenderAllowed } from "./runtime-support.js";
-export {
-  resolveDefaultGroupPolicy,
-  resolveOpenProviderRuntimeGroupPolicy,
-  warnMissingProviderGroupPolicyFallbackOnce,
-} from "./runtime-support.js";
-export { createChannelPairingController } from "./runtime-support.js";
-export { createChannelMessageReplyPipeline } from "./runtime-support.js";
-export { logTypingFailure } from "./runtime-support.js";
-export {
-  deliverTextOrMediaReply,
-  isNumericTargetId,
-  sendPayloadWithChunkedTextAndMedia,
-} from "./runtime-support.js";
-export { waitForAbortSignal } from "./runtime-support.js";
-export {
-  WEBHOOK_ANOMALY_COUNTER_DEFAULTS,
-  WEBHOOK_RATE_LIMIT_DEFAULTS,
-  applyBasicWebhookRequestGuards,
-  createFixedWindowRateLimiter,
-  createWebhookAnomalyTracker,
-  readJsonWebhookBodyOrReject,
-  registerPluginHttpRoute,
-  registerWebhookTarget,
-  registerWebhookTargetWithPluginRoute,
-  resolveWebhookPath,
-  resolveWebhookTargetWithAuthOrRejectSync,
-  withResolvedWebhookRequestPipeline,
-} from "./runtime-support.js";
-export { setZaloRuntime } from "./runtime.js";
